requirements.txt
transformers
langchain>=0.2.0
langchain-community
huggingface_hub
streamlit
python-dotenv
PyPDF2
python-docx
torch
